block: set every properties in environment
Now that we have dynamic properties, it'd be nice to set all of them
into the environment of a forked child.

To keep it simple for the moment, just upcase and prefix the key with
"BLOCK_". We might deprecate that for a consistent convention later.

This provides a simple way to pass arguments to script like this:

    [foo]
    bar=baz
    command=echo $BLOCK_BAR
    interval=2